+++ src/app/(user)/auth/login/page.tsx
@@ -1,32 +1,5 @@
-"use client";
-
-import React from "react";
-import LoginComponent from "@/components/auth/LoginComponent";
-import { useEffect } from "react";
-import { useToastStore } from "@/store/useToastStore";
-import { useSearchParams } from "next/navigation";
+import SearchParamsContent from "./SearchParamsContent";
 
 export default function LoginPage() {
-  const searchParams = useSearchParams();
-  const showToast = useToastStore((state) => state.showToast);
-
-  useEffect(() => {
-    const toastType = searchParams.get("toastType");
-    const toastMessage = searchParams.get("toastMessage");
-
-    if (toastType && toastMessage) {
-      showToast(toastMessage, toastType as any);
-    }
-  }, [searchParams, showToast]);
-
-  const styles = {
-    container: `flex items-center w-full mt-[57px]
-    tablet:w-[327px] tablet:mx-auto tablet:px-0
-    pc:w-[640px] pc:mt-[26px]`,
-  };
-  return (
-    <div className={styles.container}>
-      <LoginComponent isUser={true} />
-    </div>
-  );
+  return <SearchParamsContent />;
 }
